build: - category endpoint terminado - order endpoint no terminado - user metodos de persistencia find by usernmane y email

+++ confiteria-santana-back/src/main/java/com/vortex/infrastructure/repositories/AlergensDAO.java
@@ -33,7 +33,7 @@ public class AlergensDAO {
      * @param id the id
      * @return the alergens
      */
-    public Alergens find(int id) {
+    public Alergens find(Long id) {
         return em.find(Alergens.class, id);
     }
